Enable previewing lesson during editing
