+++ apps/platform/src/lib/infrastructure/client/pages/workspace/edit/edit-lesson-components.tsx
@@ -1,12 +1,10 @@
-import React, { useEffect, useState } from 'react';
-import { useLessonComponents } from './hooks/edit-lesson-hooks';
+import React, { useState } from 'react';
 import {
     AbortError,
     calculateMd5,
     CourseElement,
     CourseElementType,
     DefaultError,
-    DefaultLoading,
     downloadFile,
     DownloadFilesDesignerComponent,
     FormElementType,
@@ -35,7 +33,6 @@ import { TLocale } from '@maany_shr/e-class-translations';
 import { useLocale } from 'next-intl';
 import { trpc } from '../../../trpc/client';
 import { fileMetadata, shared } from '@maany_shr/e-class-models';
-import { transformLessonComponents } from '../../../utils/transform-lesson-components';
 import { generateTempId } from './utils/generate-temp-id';
 
 interface FileUploadProps {
@@ -1488,20 +1485,6 @@ export default function EditLessonComponents({
     courseVersion,
     setCourseVersion,
 }: EditLessonComponentsProps) {
-    const lessonComponentsViewModel = useLessonComponents(lessonId);
-
-    useEffect(() => {
-        if (!lessonComponentsViewModel) return;
-        if (lessonComponentsViewModel.mode !== 'default') return;
-
-        setComponents(
-            transformLessonComponents(
-                lessonComponentsViewModel.data.components,
-            ),
-        );
-        setCourseVersion(lessonComponentsViewModel.data.courseVersion);
-    }, [lessonComponentsViewModel]);
-
     const onUpClick = (id: string) => {
         setComponents((prev) => {
             const index = prev.findIndex((comp) => comp.id === id);
@@ -1533,14 +1516,6 @@ export default function EditLessonComponents({
     };
 
     const locale = useLocale() as TLocale;
-
-    if (!lessonComponentsViewModel) {
-        return <DefaultLoading locale={locale} />;
-    }
-
-    if (lessonComponentsViewModel.mode !== 'default') {
-        return <DefaultError locale={locale} />;
-    }
 
     return (
         <div className="flex flex-col gap-2">
